Move cutout_footprint to utils
